+++ .coveragerc
@@ -1,0 +1,4 @@
+[run]
+omit =
+    # omit tests
+    */_tests/*
