+++ pathy/_tests/conftest.py
@@ -7,7 +7,7 @@ from typing import Any, Generator, Optional
 
 import pytest
 
-from pathy import Pathy, use_fs, use_fs_cache, set_client_params
+from pathy import Pathy, set_client_params, use_fs, use_fs_cache
 from pathy.gcs import has_gcs
 
 has_credentials = "GCS_CREDENTIALS" in os.environ
