+++ pathy/_tests/test_cli.py
