+++ tools/clean.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+echo "Removing build files..."
+rm -rf dist/ build/ pathy.egg-info htmlcov
