+++ tools/lint.sh
@@ -4,7 +4,7 @@ set -e
 . .env/bin/activate
 
 echo "========================= mypy"
-mypy pathy
+mypy pathy --strict-equality --disallow-untyped-calls --disallow-untyped-defs
 echo "========================= flake8"
 flake8 pathy
 echo "========================= black"
