+++ tools/test.sh
@@ -3,5 +3,5 @@ set -e
 echo "Activating virtualenv... (if this fails you may need to run setup.sh first)"
 . .env/bin/activate
 echo "Running tests..."
-pytest pathy/tests --cov=pathy
+pytest pathy/_tests --cov=pathy
